Tidied item routes/controllers/templates

+++ desktop/public_html/js/app/controllers/lists/list/deleteController.js
@@ -1,10 +1,10 @@
 App.ListDeleteController = Ember.ObjectController.extend({
     actions: {
-        cancel: function ()
+        cancel : function ()
         {
             this.transitionToRoute('list', this.get('model'));
         },
-        delete: function ()
+        confirm: function ()
         {
             this.get('model').deleteRecord();
             this.get('model').save();
